+++ macros/src/lib.rs
@@ -4,13 +4,13 @@ use proc_macro::TokenStream;
 
 use syn::{DeriveInput, parse_macro_input};
 
-mod custom_mutation;
+mod custom_operation;
 
 #[proc_macro_derive(CustomOperation, attributes(seaography))]
-pub fn custom_mutation(input: TokenStream) -> TokenStream {
+pub fn custom_operation(input: TokenStream) -> TokenStream {
     let derive_input = parse_macro_input!(input as DeriveInput);
 
-    match custom_mutation::expand(derive_input) {
+    match custom_operation::expand(derive_input) {
         Ok(token_stream) => token_stream.into(),
         Err(e) => e.to_compile_error().into(),
     }
